create interface user service 👌😎👌
