+++ src/test/java/com/security/jdbc/authentication/UserRepositoryImplTest.java
@@ -1,6 +1,6 @@
 package com.security.jdbc.repository;
 
-import com.security.jdbc.authentication.repository.UserRepository;
+import com.security.jdbc.authentication.repository.UserRepositoryImpl;
 import com.security.jdbc.authentication.mapper.AuthoritiesMapper;
 import com.security.jdbc.authentication.mapper.UserInfoMapper;
 import com.security.jdbc.authentication.dto.Authorities;
@@ -24,14 +24,14 @@ import static org.mockito.ArgumentMatchers.*;
 import static org.mockito.Mockito.times;
 
 @ExtendWith(MockitoExtension.class)
-class UserRepositoryTest {
+class UserRepositoryImplTest {
 
 
     @Mock
     private JdbcTemplate jdbcTemplate;
 
     @InjectMocks
-    private UserRepository userRepository;
+    private UserRepositoryImpl userRepositoryImpl;
 
     @Test
     void addUser() {
@@ -40,7 +40,7 @@ class UserRepositoryTest {
         String password = "securePassword";
 
         // Act
-        userRepository.addUser(email, password);
+        userRepositoryImpl.addUser(email, password);
 
         // Assert
         Mockito.verify(jdbcTemplate, times(1)).update(
@@ -71,7 +71,7 @@ class UserRepositoryTest {
                         Mockito.eq(email)
                         )).thenReturn(userInfo);
 
-        UserInfo result = userRepository.findOneUserByEnabledAndEmail(email);
+        UserInfo result = userRepositoryImpl.findOneUserByEnabledAndEmail(email);
 
         // Assert
         assertNotNull(result);
@@ -95,7 +95,7 @@ class UserRepositoryTest {
         )).thenThrow(new EmptyResultDataAccessException(1));
 
         // Act
-        UserInfo result = userRepository.findOneUserByEnabledAndEmail(email);
+        UserInfo result = userRepositoryImpl.findOneUserByEnabledAndEmail(email);
 
         // Assert
         assertNull(result);
@@ -122,7 +122,7 @@ class UserRepositoryTest {
         )).thenReturn(mockAuthorities);
 
         // Act
-        List<GrantedAuthority> result = userRepository.getUserAuthoritiesByUserEmail(email);
+        List<GrantedAuthority> result = userRepositoryImpl.getUserAuthoritiesByUserEmail(email);
 
         // Assert
         assertNotNull(result);
@@ -159,7 +159,7 @@ class UserRepositoryTest {
         )).thenReturn(mockUsers);
 
         // Act
-        List<UserInfo> result = userRepository.getAllUsers();
+        List<UserInfo> result = userRepositoryImpl.getAllUsers();
 
         // Assert
         assertNotNull(result);
