+++ src/test/java/com/security/jdbc/authentication/UserServiceImplTest.java
@@ -2,7 +2,7 @@ package com.security.jdbc.service;
 
 
 import com.security.jdbc.authentication.service.UserService;
-import com.security.jdbc.authentication.repository.UserRepository;
+import com.security.jdbc.authentication.repository.UserRepositoryImpl;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -16,7 +16,7 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 public class UserServiceTest {
 
     @Mock
-    private UserRepository userRepository;
+    private UserRepositoryImpl userRepositoryImpl;
 
     @Mock
     private PasswordEncoder passwordEncoder;
@@ -37,7 +37,7 @@ public class UserServiceTest {
         var result = userService.addUser(email, password);
 
         Mockito.verify(passwordEncoder, Mockito.times(1)).encode(password);
-        Mockito.verify(userRepository, Mockito.times(1)).addUser(email, "encodedPassword"); //must be encoded. See the service which repository.add saves email and encoded_password
+        Mockito.verify(userRepositoryImpl, Mockito.times(1)).addUser(email, "encodedPassword"); //must be encoded. See the service which repository.add saves email and encoded_password
 
         Assertions.assertEquals("User registered successfully!", result);
     }
